refactor: move relevance tokenisation into query
prompt imported index only for RelevanceTerms, which blocked index from using
prompt.Terms. The tokeniser and its CJK helpers are pure, so they move to the
leaf package and index keeps thin wrappers.
